Struct gstreamer_sys::GstMessageType

#[repr(C)]

pub struct GstMessageType { /* fields omitted */ }

Methods

impl GstMessageType

const UNKNOWN: GstMessageType

UNKNOWN: GstMessageType = GstMessageType{bits: 0,}

const EOS: GstMessageType

EOS: GstMessageType = GstMessageType{bits: 1,}

const ERROR: GstMessageType

ERROR: GstMessageType = GstMessageType{bits: 2,}

const WARNING: GstMessageType

WARNING: GstMessageType = GstMessageType{bits: 4,}

const INFO: GstMessageType

INFO: GstMessageType = GstMessageType{bits: 8,}

const TAG: GstMessageType

TAG: GstMessageType = GstMessageType{bits: 16,}

const BUFFERING: GstMessageType

BUFFERING: GstMessageType = GstMessageType{bits: 32,}

const STATE_CHANGED: GstMessageType

STATE_CHANGED: GstMessageType = GstMessageType{bits: 64,}

const STATE_DIRTY: GstMessageType

STATE_DIRTY: GstMessageType = GstMessageType{bits: 128,}

const STEP_DONE: GstMessageType

STEP_DONE: GstMessageType = GstMessageType{bits: 256,}

const CLOCK_PROVIDE: GstMessageType

CLOCK_PROVIDE: GstMessageType = GstMessageType{bits: 512,}

const CLOCK_LOST: GstMessageType

CLOCK_LOST: GstMessageType = GstMessageType{bits: 1024,}

const NEW_CLOCK: GstMessageType

NEW_CLOCK: GstMessageType = GstMessageType{bits: 2048,}

const STRUCTURE_CHANGE: GstMessageType

STRUCTURE_CHANGE: GstMessageType = GstMessageType{bits: 4096,}

const STREAM_STATUS: GstMessageType

STREAM_STATUS: GstMessageType = GstMessageType{bits: 8192,}

const APPLICATION: GstMessageType

APPLICATION: GstMessageType = GstMessageType{bits: 16384,}

const ELEMENT: GstMessageType

ELEMENT: GstMessageType = GstMessageType{bits: 32768,}

const SEGMENT_START: GstMessageType

SEGMENT_START: GstMessageType = GstMessageType{bits: 65536,}

const SEGMENT_DONE: GstMessageType

SEGMENT_DONE: GstMessageType = GstMessageType{bits: 131072,}

const DURATION_CHANGED: GstMessageType

DURATION_CHANGED: GstMessageType = GstMessageType{bits: 262144,}

const LATENCY: GstMessageType

LATENCY: GstMessageType = GstMessageType{bits: 524288,}

const ASYNC_START: GstMessageType

ASYNC_START: GstMessageType = GstMessageType{bits: 1048576,}

const ASYNC_DONE: GstMessageType

ASYNC_DONE: GstMessageType = GstMessageType{bits: 2097152,}

const REQUEST_STATE: GstMessageType

REQUEST_STATE: GstMessageType = GstMessageType{bits: 4194304,}

const STEP_START: GstMessageType

STEP_START: GstMessageType = GstMessageType{bits: 8388608,}

const QOS: GstMessageType

QOS: GstMessageType = GstMessageType{bits: 16777216,}

const PROGRESS: GstMessageType

PROGRESS: GstMessageType = GstMessageType{bits: 33554432,}

const TOC: GstMessageType

TOC: GstMessageType = GstMessageType{bits: 67108864,}

const RESET_TIME: GstMessageType

RESET_TIME: GstMessageType = GstMessageType{bits: 134217728,}

const STREAM_START: GstMessageType

STREAM_START: GstMessageType = GstMessageType{bits: 268435456,}

const NEED_CONTEXT: GstMessageType

NEED_CONTEXT: GstMessageType = GstMessageType{bits: 536870912,}

const HAVE_CONTEXT: GstMessageType

HAVE_CONTEXT: GstMessageType = GstMessageType{bits: 1073741824,}

const EXTENDED: GstMessageType

EXTENDED: GstMessageType = GstMessageType{bits: 2147483648,}

const DEVICE_ADDED: GstMessageType

DEVICE_ADDED: GstMessageType = GstMessageType{bits: 2147483649,}

const DEVICE_REMOVED: GstMessageType

DEVICE_REMOVED: GstMessageType = GstMessageType{bits: 2147483650,}

const PROPERTY_NOTIFY: GstMessageType

PROPERTY_NOTIFY: GstMessageType = GstMessageType{bits: 2147483651,}

const STREAM_COLLECTION: GstMessageType

STREAM_COLLECTION: GstMessageType = GstMessageType{bits: 2147483652,}

const STREAMS_SELECTED: GstMessageType

STREAMS_SELECTED: GstMessageType = GstMessageType{bits: 2147483653,}

const REDIRECT: GstMessageType

REDIRECT: GstMessageType = GstMessageType{bits: 2147483654,}

const ANY: GstMessageType

ANY: GstMessageType = GstMessageType{bits: 4294967295,}

pub fn empty() -> GstMessageType

Returns an empty set of flags.

pub fn all() -> GstMessageType

Returns the set containing all flags.

pub fn bits(&self) -> c_uint

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: c_uint) -> Option<GstMessageType>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub fn from_bits_truncate(bits: c_uint) -> GstMessageType

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub fn intersects(&self, other: GstMessageType) -> bool

Returns true if there are flags common to both self and other.

pub fn contains(&self, other: GstMessageType) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: GstMessageType)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: GstMessageType)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: GstMessageType)

Toggles the specified flags in-place.

pub fn set(&mut self, other: GstMessageType, value: bool)

Inserts or removes the specified flags depending on the passed value.